+++ docs/REQUIREMENTS.md
@@ -14,4 +14,4 @@
 - [x] 백엔드 서버 선택 기능 구현
 - [x] 백엔드 서버 연결
 - [x] 상품 아이디와 장바구니 아이디를 분리
-- [ ] 반응형 (모바일, PC)
+- [x] 반응형 (모바일, PC)
