+++ package.json
@@ -70,6 +70,7 @@
     "gh-pages": "^5.0.0",
     "msw": "^1.2.1",
     "prop-types": "^15.8.1",
+    "react-error-boundary": "^4.0.8",
     "react-router-dom": "^6.11.1",
     "recoil": "^0.7.7",
     "storybook": "^7.0.10",
