+++ src/components/mainPage/productCardGrid/ProductCardGrid.tsx
@@ -1,13 +1,12 @@
 import styled from 'styled-components';
-import { Product } from '../../../types/Product';
 
 import { ProductCard } from '../productCard/ProductCard';
+import { productItemsState } from '../../../recoil/atoms/productItemsAtom';
+import { useRecoilValue } from 'recoil';
 
-interface ProductCardListProps {
-  products: Product[];
-}
+export const ProductCardGrid = () => {
+  const products = useRecoilValue(productItemsState);
 
-export const ProductCardGrid = ({ products }: ProductCardListProps) => {
   return (
     <Style.Container>
       {products.map((product) => (
