+++ src/constants/errorMessage.ts
@@ -1,0 +1,6 @@
+export const errorMessage = Object.freeze({
+  PRODUCT_ITEMS_FETCH_ERROR: '상품을 불러오지 못했어요',
+  CART_ITEMS_FETCH_ERROR: '장바구니를 불러오지 못했어요',
+  ORDER_ITEMS_FETCH_ERROR: '주문 목록을 불러오지 못했어요',
+  ORDER_DETAIL_FETCH_ERROR: '주문 내역 상세를 불러오지 못했어요',
+});
