+++ src/pages/Main.tsx
@@ -1,25 +1,19 @@
+import { ErrorBoundary } from 'react-error-boundary';
 import { ProductCardGrid } from '../components/mainPage/productCardGrid/ProductCardGrid';
 import { Layout } from '../layout';
-import { useEffect, useState } from 'react';
-import { Product } from '../types/Product';
-import { useProductFetch } from '../hooks/fetch/useProductFetch';
-import { useRecoilValue } from 'recoil';
-import { APIAtom } from '../recoil/atoms/serverAtom';
+import { Fallback } from '../components/error/Fallback';
+import { errorMessage } from '../constants/errorMessage';
 
 function Main() {
-  const [products, setProducts] = useState<Product[]>();
-  const { getProductList } = useProductFetch();
-  const apiEndPoint = useRecoilValue(APIAtom);
-
-  useEffect(() => {
-    getProductList().then((data) => {
-      setProducts(data);
-    });
-  }, [apiEndPoint]);
-
   return (
     <Layout>
-      {products ? <ProductCardGrid products={products} /> : <>loading...</>}
+      <ErrorBoundary
+        FallbackComponent={() => (
+          <Fallback error={new Error(errorMessage.PRODUCT_ITEMS_FETCH_ERROR)} />
+        )}
+      >
+        <ProductCardGrid />
+      </ErrorBoundary>
     </Layout>
   );
 }
