+++ src/recoil/atoms/cartAtom.ts
@@ -2,28 +2,27 @@ import { atom } from 'recoil';
 import { Product } from '../../types/Product';
 import { base64 } from '../../constants/user';
 
-export interface CartProductDetail {
+export interface CartItemDetail {
   id: number;
   quantity: number;
   product: Product;
 }
 
-export const cartItemsState = atom<CartProductDetail[]>({
+export const cartItemsState = atom<CartItemDetail[]>({
   key: 'cartItemsState',
   default: [],
   effects: [
     ({ setSelf, trigger }) => {
-      const getCartItems = async () => {
-        const response = await fetch(`/cart-items`, {
+      const getCartItems = () => {
+        const response = fetch(`/cart-items`, {
           method: 'GET',
           headers: {
             Authorization: `Basic ${base64}`,
             'Content-Type': 'application/json',
           },
-        });
-        const cartItems = await response.json();
+        }).then((res) => res.json() as Promise<CartItemDetail[]>);
 
-        setSelf(cartItems);
+        setSelf(response);
       };
 
       if (trigger === 'get') getCartItems();
