+++ src/recoil/selectors/priceSummarySelector.ts
@@ -68,12 +68,3 @@ export const priceSummaryState = selector({
     };
   },
 });
-
-// 사용 가능한 포인트 구하기
-// 1. 선택된 상품들 중 포인트를 사용 가능한 상품의 가격을 모두 더해 총 사용 가능한 포인트를 계산한다.
-// 2. 유저가 가진 포인트를 가져와 이보다 1번의 포인트가 더 크다면 유저가 가진 포인트를, 만약 유저가 가진 포인트가 더 크다면 1번의 포인트를 반환한다.
-
-// 적립 가능한 포인트 구하기
-// 1. 선택된 상품들의 수량 * 가격에 각 상품의 포인트 전환 비율을 곱하여 모두 더한 값을 반환한다.
-
-//
